refactor(preferences): extract useUpdatePreferences hook
