+++ src/components/features/my-filters/categories/uncategorized-heading-dropdown.tsx
@@ -3,7 +3,6 @@
 import { useState } from "react";
 import { api } from "@/trpc/react";
 import { ArrowDownToLine, ArrowUpToLine, EllipsisIcon, Share } from "lucide-react";
-import { toast } from "sonner";
 
 import { useGetUserFiltersByCategory } from "@/hooks/use-get-user-filters-by-category";
 import { Button } from "@/components/ui/button";
@@ -15,6 +14,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { ShareWithUserDialog } from "@/components/features/my-filters/shared-filters/share-with-user-dialog";
+import { useUpdatePreferences } from "@/components/features/my-filters/hooks/use-update-preferences";
 
 interface UncategorizedHeadingDropdownProps {
   categoryId: number | null;
@@ -28,28 +28,8 @@ export function UncategorizedHeadingDropdown({
 
   const hasFilters = uncategorizedFilters.length > 0;
 
-  const utils = api.useUtils();
   const preferencesQuery = api.userPreferences.get.useQuery();
-  const updatePreferences = api.userPreferences.update.useMutation({
-    onMutate: async ({ uncategorizedPosition }) => {
-      await utils.userPreferences.get.cancel();
-      const previous = utils.userPreferences.get.getData();
-      utils.userPreferences.get.setData(undefined, (old) => ({
-        userId: old?.userId ?? "",
-        uncategorizedPosition,
-      }));
-      return { previous };
-    },
-    onError: (err, _, context) => {
-      if (context?.previous) {
-        utils.userPreferences.get.setData(undefined, context.previous);
-      }
-      toast.error(err.message || "Failed to update preference");
-    },
-    onSettled: () => {
-      utils.userPreferences.get.invalidate();
-    },
-  });
+  const updatePreferences = useUpdatePreferences();
 
   const currentPosition = preferencesQuery.data?.uncategorizedPosition ?? "top";
   const togglePosition = currentPosition === "top" ? "bottom" : "top";
